@@ -1,16 +1,67 @@
 {
-  "name": "my_react_app",
-  "version": "0.1.0",
+  "name": "material-dashboard-pro-react",
+  "version": "1.9.0",
   "private": true,
   "dependencies": {
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2",
-    "react-scripts": "1.1.5"
+    "@material-ui/core": "4.9.14",
+    "@material-ui/icons": "4.9.1",
+    "aws-sdk": "^2.889.0",
+    "chartist": "0.10.1",
+    "classnames": "2.2.6",
+    "dynamodb": "^1.3.0",
+    "history": "4.10.1",
+    "match-sorter": "4.1.0",
+    "moment": "2.26.0",
+    "mqtt": "^4.2.6",
+    "node-sass": "4.14.1",
+    "nouislider": "14.5.0",
+    "perfect-scrollbar": "1.5.0",
+    "react": "16.13.1",
+    "react-big-calendar": "0.24.6",
+    "react-bootstrap-sweetalert": "5.1.9",
+    "react-chartist": "0.14.3",
+    "react-datetime": "2.16.3",
+    "react-dom": "16.13.1",
+    "react-google-maps": "9.4.5",
+    "react-jvectormap": "0.0.16",
+    "react-router-dom": "5.2.0",
+    "react-scripts": "3.4.1",
+    "react-swipeable-views": "0.13.9",
+    "react-table": "7.1.0",
+    "react-tagsinput": "3.19.0"
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "react-scripts build",
+    "build": "react-scripts build && gulp licenses",
     "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "deploy": "npm run build",
+    "lint:check": "eslint . --ext=js,jsx;  exit 0",
+    "lint:fix": "eslint . --ext=js,jsx --fix;  exit 0",
+    "install:clean": "rm -rf node_modules/ && rm -rf package-lock.json && npm install && npm start",
+    "compile-sass": "node-sass src/assets/scss/material-dashboard-pro-react.scss src/assets/css/material-dashboard-pro-react.css",
+    "minify-sass": "node-sass src/assets/scss/material-dashboard-pro-react.scss src/assets/css/material-dashboard-pro-react.min.css --output-style compressed",
+    "map-sass": "node-sass src/assets/scss/material-dashboard-pro-react.scss src/assets/css/material-dashboard-pro-react.css --source-map true"
+  },
+  "optionalDependencies": {
+    "@types/googlemaps": "3.39.6",
+    "@types/markerclustererplus": "2.1.33",
+    "ajv": "6.12.2",
+    "typescript": "3.9.3"
+  },
+  "devDependencies": {
+    "eslint-config-prettier": "6.11.0",
+    "eslint-plugin-prettier": "3.1.3",
+    "gulp": "4.0.2",
+    "gulp-append-prepend": "1.0.8",
+    "prettier": "2.0.5"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": []
   }
 }
